document: input_data put
