+++ tests/data/input_data/lenna.zip
