+++ tests/test_main.py
@@ -192,132 +192,6 @@ class TestFilesystem:
         )
 
 
-class TestInputData:
-    command_name = "input_data"
-
-    @classmethod
-    def setup_class(cls):
-        annofab_service = annofabapi.build()
-        task, _ = annofab_service.api.get_task(project_id, task_id)
-        cls.input_data_id = task["input_data_id_list"][0]
-
-    def test_delete_input_data(self):
-        main(["input_data", "delete", "--project_id", project_id, "--input_data_id", "foo", "--yes"])
-
-    def test_list_input_data(self):
-        out_file = str(out_path / "input_data.json")
-        main(
-            [
-                "input_data",
-                "list",
-                "--project_id",
-                project_id,
-                "--input_data_query",
-                '{"input_data_name": "abcdefg"}',
-                "--add_details",
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_input_data_merged_task_with_downloading(self):
-        out_file = str(out_path / "input_data.csv")
-        main(
-            [
-                "input_data",
-                "list_merged_task",
-                "--project_id",
-                project_id,
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_input_data_merged_task_with_json(self):
-        out_file = str(out_path / "input_data.csv")
-        main(
-            [
-                "input_data",
-                "list_merged_task",
-                "--task_json",
-                str(data_path / "task.json"),
-                "--input_data_json",
-                str(data_path / "input_data.json"),
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_list_input_data_with_json(self):
-        out_file = str(out_path / "input_data.csv")
-        main(
-            [
-                "input_data",
-                "list_with_json",
-                "--project_id",
-                project_id,
-                "--input_data_query",
-                '{"input_data_name": "abcdefg"}',
-                "--input_data_id",
-                "test1",
-                "test2",
-                "--output",
-                out_file,
-            ]
-        )
-
-    def test_put_input_data(self):
-        csv_file = str(data_path / "input_data2.csv")
-        main(
-            [
-                self.command_name,
-                "put",
-                "--project_id",
-                project_id,
-                "--csv",
-                csv_file,
-                "--overwrite",
-                "--yes",
-                "--parallelism",
-                "2",
-            ]
-        )
-
-    def test_update_metadata(self):
-        main(
-            [
-                self.command_name,
-                "update_metadata",
-                "--project_id",
-                project_id,
-                "--input_data_id",
-                self.input_data_id,
-                "--metadata",
-                '{"attr1":"foo"}',
-                "--yes",
-            ]
-        )
-
-    @pytest.mark.submitting_job
-    def test_put_input_data_with_zip(self):
-        # 注意：ジョブ登録される
-        zip_file = str(data_path / "lenna.zip")
-        main(
-            [
-                self.command_name,
-                "put",
-                "--project_id",
-                project_id,
-                "--zip",
-                zip_file,
-                "--wait",
-                "--yes",
-                "--wait_options",
-                '{"interval":1, "max_tries":1}',
-            ]
-        )
-
-
 class TestInspectionComment:
     def test_list_inspection_comment(self):
         out_file = str(out_path / "inspection_comment.csv")
